+++ huspacy/huspacy/components/__init__.py
@@ -3,3 +3,4 @@ from . import sentencizer
 from . import lemma_postprocessing
 from . import lookup_lemmatizer
 from . import sentiment
+from . import nerpp